Separate logic from DOM manipulation
