+++ src/selectors.js
@@ -1,0 +1,9 @@
+/* eslint-disable no-unused-vars */
+export const cityInput = document.querySelector('#city');
+export const cityLabel = document.querySelector('#city-label');
+export const errorDiv = document.querySelector('#error');
+export const feelsDiv = document.querySelector('#feels');
+export const humidityDiv = document.querySelector('#humidity');
+export const tempDiv = document.querySelector('#temp');
+export const url = 'http://api.openweathermap.org/data/2.5/weather?appid=b93ac565c07c898f8ab078f813afa920&units=metric';
+export const windsDiv = document.querySelector('#winds');
